continue refactor, get runner code working
