+++ dbt/task/run.py
@@ -16,8 +16,8 @@ class RunTask:
     def compile(self):
         compiler = Compiler(self.project, BaseCreateTemplate)
         compiler.initialize()
-        created_models, created_analyses = compiler.compile()
-        print("Compiled {} models and {} analyses".format(created_models, created_analyses))
+        created_models, created_tests, created_analyses = compiler.compile()
+        print("Compiled {} models, {} tests, and {} analyses".format(created_models, created_tests, created_analyses))
 
     def run(self):
         self.compile()
